load-tweets/requirements.txt
google-cloud-bigquery==3.3.2
tweepy==4.10.1
python-dotenv==0.21.0
